# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QilbeeDB Platform API",
    "version": "0.4.0",
    "description": "Durable identity, scoped memory with model-bound semantic vector retrieval, evidence-driven procedural learning and learned-tool development receipts for AI agents."
  },
  "servers": [
    {
      "url": "http://localhost:7474",
      "description": "Local Docker or development server"
    }
  ],
  "security": [
    {
      "platformCredential": []
    }
  ],
...
  "paths": {
...
    "/api/v1/identity": {
      "get": {
        "operationId": "whoAmI",
        "summary": "Inspect the current durable credential",
        "description": "",
        "tags": [
          "Identity"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or revoked credential",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure or explicit stored-version inconsistency",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/credentials": {
      "post": {
        "operationId": "issueCredential",
        "summary": "Issue a credential in the authenticated tenant",
        "description": "Requires credential_admin. Secret delivery is one-time and not idempotent.",
        "tags": [
          "Identity"
        ],
        "responses": {
          "201": {
            "description": "Successful operation",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedCredential"
                }
              }
            }
          },
          "400": {
            "description": "Invalid fields or unsupported contract version",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or revoked credential",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Capability, tenant or scope is not granted",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency payload or expected revision conflict",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Request exceeds 65,536 bytes",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure or explicit stored-version inconsistency",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/credentials/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "inspectCredential",
        "summary": "Inspect sanitized credential history",
        "description": "Requires credential_admin in the same tenant.",
        "tags": [
          "Identity"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid fields or unsupported contract version",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or revoked credential",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Capability, tenant or scope is not granted",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure or explicit stored-version inconsistency",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "platformCredential": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "qdb1 opaque credential",
        "description": "Provision through the trusted local bootstrap command and authenticated credential API."
      }
    },
    "schemas": {
...
      "CredentialResponse": {
        "type": "object",
        "properties": {
          "contract_version": {
            "type": "integer",
            "const": 1
          },
          "credential": {
            "$ref": "#/components/schemas/Credential"
          }
        },
        "required": [
          "contract_version",
          "credential"
        ],
        "additionalProperties": false
      },
      "IssuedCredential": {
        "type": "object",
        "properties": {
          "contract_version": {
            "type": "integer",
            "const": 1
          },
          "credential": {
            "$ref": "#/components/schemas/Credential"
          },
          "secret": {
            "type": "string",
            "pattern": "^qdb1_[0-9a-f]{32}_[A-Za-z0-9_-]{43}$",
            "description": "Returned once. Store securely; never log."
          }
        },
        "required": [
          "contract_version",
          "credential",
          "secret"
        ],
        "additionalProperties": false
      },
      "IssueRequest": {
        "type": "object",
        "properties": {
          "contract_version": {
            "type": "integer",
            "const": 1
          },
          "spec": {
            "$ref": "#/components/schemas/CredentialSpec"
          }
        },
        "required": [
          "contract_version",
          "spec"
        ],
        "additionalProperties": false,
        "examples": [
          {
            "contract_version": 1,
            "spec": {
              "subject_id": "research-agent-user",
              "capabilities": [
                "memory_read",
                "memory_write",
                "procedure_propose"
              ],
              "grants": [
                {
                  "project_id": "research",
                  "mission_id": "mission-42",
                  "agent_id": "researcher",
                  "visibility": "shared"
                }
              ],
              "expires_at_millis": null
            }
          }
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "contract_version": {
            "type": "integer",
            "const": 1
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "unauthorized",
                  "forbidden",
                  "not_found",
                  "method_not_allowed",
                  "invalid_request",
                  "unsupported_contract_version",
                  "revision_conflict",
                  "record_not_found",
                  "idempotency_conflict",
                  "storage_inconsistency",
                  "internal_error"
                ]
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "contract_version",
          "error"
        ],
        "additionalProperties": false
      },
...
    }
  }
}